checkpoint for studydesign module
