+++ study/src/org/labkey/study/view/BaseStudyPage.java
@@ -70,16 +70,6 @@ public abstract class BaseStudyPage extends JspBase
         return getStudy().getAssaySpecimenConfigs();
     }
 
-    protected List<ProductImpl> getStudyProducts(User user, String role)
-    {
-        return getStudy().getStudyProducts(user, role);
-    }
-
-    protected List<TreatmentImpl> getStudyTreatments(User user)
-    {
-        return getStudy().getStudyTreatments(user);
-    }
-
     protected WikiRendererType[] getRendererTypes()
     {
         return WikiRendererType.values();
