+++ studydesign/module.properties
@@ -1,0 +1,5 @@
+ModuleClass: org.labkey.studydesign.StudyDesignModule
+License: Apache 2.0
+LicenseURL: http://www.apache.org/licenses/LICENSE-2.0
+SupportedDatabases: mssql, pgsql
+ManageVersion: true
