+++ studydesign/src/org/labkey/studydesign/view/StudyDesignConfigureMenuItem.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2017 LabKey Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.labkey.studydesign.view;
+
+import org.labkey.api.data.Container;
+import org.labkey.api.query.QueryUrls;
+import org.labkey.api.util.PageFlowUtil;
+import org.labkey.api.view.ActionURL;
+import org.labkey.api.view.NavTree;
+
+public class StudyDesignConfigureMenuItem extends NavTree
+{
+    public StudyDesignConfigureMenuItem(String text, String schemaName, String queryName, Container container)
+    {
+        super(text);
+
+        ActionURL url = new ActionURL();
+        url.setContainer(container);
+        url.addParameter("schemaName", schemaName);
+        url.addParameter("query.queryName", queryName);
+        setHref(PageFlowUtil.urlProvider(QueryUrls.class).urlExecuteQuery(url).toString());
+        setTarget("_blank");  // issue 19493
+    }
+}
